moved components into proper folder
