+++ mobile/App.tsx
@@ -2,8 +2,8 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, View } from 'react-native'
 import { getApp } from 'firebase/app';
 import { auth } from '../mobile/firebase/firebaseConfig.js';
-import PhoneAuth from './components/CustomButton/PhoneAuth';
-import DataEntry from './components/CustomButton/DataEntry';
+import PhoneAuth from './components/PhoneAuth';
+import DataEntry from './components/DataEntry';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
 
 export default function App() {
