+++ mobile/components/DataEntry.tsx
